requirements.txt
asgiref==3.3.1
bleach==3.2.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.0.11
django-filter==2.4.0
django-reversion==3.0.9
djangorestframework==3.12.2
docutils==0.16
idna==2.10
keyring==22.0.0
packaging==20.8
pkginfo==1.7.0
Pygments==2.7.4
pyparsing==2.4.7
pytz==2020.5
readme-renderer==28.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
segno==1.3.3
six==1.15.0
sqlparse==0.4.1
tqdm==4.56.0
twine==3.3.0
urllib3==1.26.2
webencodings==0.5.1
